[client] 仕様書項目新規作成＆編集における商品取得処理をhooks化
